+++ commercetools/commercetools-sdk-java-api/src/main/java-generated/com/commercetools/api/models/message/OrderLineItemDiscountSetMessageBuilder.java
@@ -64,7 +64,7 @@ public class OrderLineItemDiscountSetMessageBuilder implements Builder<OrderLine
 
     private java.util.List<com.commercetools.api.models.cart.DiscountedLineItemPriceForQuantity> discountedPricePerQuantity;
 
-    private com.commercetools.api.models.common.Money totalPrice;
+    private com.commercetools.api.models.common.CentPrecisionMoney totalPrice;
 
     @Nullable
     private com.commercetools.api.models.cart.TaxedItemPrice taxedPrice;
@@ -394,8 +394,8 @@ public class OrderLineItemDiscountSetMessageBuilder implements Builder<OrderLine
      */
 
     public OrderLineItemDiscountSetMessageBuilder totalPrice(
-            Function<com.commercetools.api.models.common.MoneyBuilder, com.commercetools.api.models.common.MoneyBuilder> builder) {
-        this.totalPrice = builder.apply(com.commercetools.api.models.common.MoneyBuilder.of()).build();
+            Function<com.commercetools.api.models.common.CentPrecisionMoneyBuilder, com.commercetools.api.models.common.CentPrecisionMoneyBuilder> builder) {
+        this.totalPrice = builder.apply(com.commercetools.api.models.common.CentPrecisionMoneyBuilder.of()).build();
         return this;
     }
 
@@ -406,8 +406,8 @@ public class OrderLineItemDiscountSetMessageBuilder implements Builder<OrderLine
      */
 
     public OrderLineItemDiscountSetMessageBuilder withTotalPrice(
-            Function<com.commercetools.api.models.common.MoneyBuilder, com.commercetools.api.models.common.Money> builder) {
-        this.totalPrice = builder.apply(com.commercetools.api.models.common.MoneyBuilder.of());
+            Function<com.commercetools.api.models.common.CentPrecisionMoneyBuilder, com.commercetools.api.models.common.CentPrecisionMoney> builder) {
+        this.totalPrice = builder.apply(com.commercetools.api.models.common.CentPrecisionMoneyBuilder.of());
         return this;
     }
 
@@ -418,7 +418,7 @@ public class OrderLineItemDiscountSetMessageBuilder implements Builder<OrderLine
      */
 
     public OrderLineItemDiscountSetMessageBuilder totalPrice(
-            final com.commercetools.api.models.common.Money totalPrice) {
+            final com.commercetools.api.models.common.CentPrecisionMoney totalPrice) {
         this.totalPrice = totalPrice;
         return this;
     }
@@ -676,7 +676,7 @@ public class OrderLineItemDiscountSetMessageBuilder implements Builder<OrderLine
      * @return totalPrice
      */
 
-    public com.commercetools.api.models.common.Money getTotalPrice() {
+    public com.commercetools.api.models.common.CentPrecisionMoney getTotalPrice() {
         return this.totalPrice;
     }